+++ src/lib.rs
@@ -284,7 +284,7 @@ fn fft_result_to_spectrum(
         // ### BEGIN filtering: results in lower calculation and memory overhead!
         // check lower bound frequency (inclusive)
         .filter(|(fr, _fft_result)| {
-            maybe_min.map_or(true, |min_fr| {
+            maybe_min.is_none_or(|min_fr| {
                 // inclusive!
                 // attention: due to the frequency resolution, we do not necessarily hit
                 //            exactly the frequency, that a user requested
@@ -294,7 +294,7 @@ fn fft_result_to_spectrum(
         })
         // check upper bound frequency (inclusive)
         .filter(|(fr, _fft_result)| {
-            maybe_max.map_or(true, |max_fr| {
+            maybe_max.is_none_or(|max_fr| {
                 // inclusive!
                 // attention: due to the frequency resolution, we do not necessarily hit
                 //            exactly the frequency, that a user requested